# openapi.yaml
openapi: '3.0.0'
info:
  title: Audio Transcription API
  description: API for transcribing audio files using OpenAI's Whisper API
  version: '1.0.0'

servers:
  - url: http://localhost:3000/api/v1
    description: Development server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Transcription:
      type: object
      properties:
        id:
          type: integer
          format: int64
        text:
          type: string
        fileName:
          type: string
        duration:
          type: number
          format: float
        createdAt:
          type: string
          format: date-time

    Error:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string

    LoginDto:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    TokenResponse:
      type: object
      properties:
        access_token:
          type: string

security:
  - bearerAuth: []

paths:
  /auth/register:
    post:
      security: []
      tags:
        - auth
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'

  /auth/login:
    post:
      security: []
      tags:
        - auth
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'

  /transcriptions/upload:
    post:
      tags:
        - transcriptions
      summary: Upload audio file for transcription
      description: Transcribes an audio file using OpenAI's Whisper API
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                audio:
                  type: string
                  format: binary
                  description: Audio file to transcribe (mp3, wav, m4a, or ogg). Max size 25MB.
      responses:
        '201':
          description: Audio file successfully transcribed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transcription'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /transcriptions:
    get:
      tags:
        - transcriptions
      summary: List all transcriptions
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Items per page
        - in: query
          name: search
          schema:
            type: string
          description: Search text in transcriptions
      responses:
        '200':
          description: List of transcriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transcription'

  /transcriptions/{id}:
    get:
      tags:
        - transcriptions
      summary: Get transcription by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Transcription found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transcription'
        '404':
          description: Transcription not found

    delete:
      tags:
        - transcriptions
      summary: Delete transcription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Transcription deleted
        '404':
          description: Transcription not found

  /health:
    get:
      tags:
        - health
      summary: Check API health
      responses:
        '200':
          description: Health check results
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  details:
                    type: object